ai-service/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-multipart==0.0.20
qdrant-client==1.13.0
google-generativeai==0.8.4
httpx==0.28.1
pypdf==5.1.0
python-docx==1.1.2
numpy==2.2.1
pydantic-settings==2.7.1
google-genai>=1.0.0